fix: address TanStack Start compatibility feedback
