+++ crates/perry-hir/src/lower/lower_module_fn.rs
@@ -14,7 +14,9 @@ use swc_ecma_ast as ast;
 
 use super::*;
 use crate::ir::*;
-use crate::lower_types::{infer_hoisted_text_codec_var_type, require_literal_specifier};
+use crate::lower_types::hoisted_text_codec::{
+    infer_hoisted_text_codec_var_type, require_literal_specifier,
+};
 
 fn module_has_strict_mode(ast_module: &ast::Module, source_file_path: &str) -> bool {
     // A file is strict-mode code exactly when Node runs it as an ES module. Three
